refactor: 컴포넌트 분리 - 컴ㅍ넌트 분리 - Type 지정
