+++ src/app/(auth)/(with-header)/questions/create/index.ts
@@ -1,0 +1,2 @@
+export * from "./questionCreate.type";
+export * from "./mocks/questoinCreateData";
