+++ src/app/(auth)/(with-header)/questions/create/mocks/questoinCreateData.ts
@@ -1,0 +1,22 @@
+export const tagConditions = [
+  { value: "designer", txt: "디자이너" },
+  { value: "developer", txt: "개발자" },
+  { value: "planner", txt: "기획자" },
+  { value: "publisher", txt: "퍼블리셔" },
+];
+
+export const tags = [
+  { value: "react", txt: "React" },
+  { value: "js", txt: "JavaScript" },
+  { value: "next", txt: "Next" },
+  { value: "node", txt: "Node" },
+  { value: "php", txt: "PHP" },
+  { value: "java", txt: "JAVA" },
+  { value: "python", txt: "Python" },
+];
+
+export const level = [
+  { value: "easy", txt: "상" },
+  { value: "medium", txt: "중" },
+  { value: "high", txt: "하" },
+];
